requirements.txt
fastapi
uvicorn
pytest
python-dotenv
requests
mcp[cli]
httpx 